+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+## [next]
+
 ## 0.21.0
 - Adding AssetsCache.readBinaryFile
 - Splitting debugMode from recordFps mode