+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Adding support for multi touch tap and drag events
 - Fix animations example
 - Add possibility for infinite and alternating effects
+- Add rotational effect for PositionComponents
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice