+++ CHANGELOG.md
@@ -4,6 +4,8 @@
  - Fixing BaseGame tap detectors issues
  - Adding FlameSpriteWidget
  - Adding FlameAnimationWidget
+ - Upgrade Flutter SVG to fix for flame web
+ - Add linting to all the examples
 
 ## 0.21.0
 - Adding AssetsCache.readBinaryFile
@@ -11,6 +13,7 @@
 - Adding support for multi touch tap and drag events
 - Fix animations example
 - Add possibility for infinite and alternating effects
+- Add rotational effect for PositionComponents
 
 ## 0.20.2
 - Fix text component bug with anchor being applied twice