+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - Fixing BaseGame tap detectors issues
  - Upgrade Flutter SVG to fix for flame web
  - Add linting to all the examples
+ - Run linting only on affected and changed examples
 
 ## 0.21.0
 - Adding AssetsCache.readBinaryFile