+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Upgrade Flutter SVG to fix for flame web
  - Add linting to all the examples
  - Run linting only on affected and changed examples
+ - Add SequenceEffect
 
 ## 0.21.0
 - Adding AssetsCache.readBinaryFile