+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## [next]
+## 0.22.0
  - Fixing BaseGame tap detectors issues
  - Adding SpriteWidget
  - Adding AnimationWidget